requirements.txt
fastapi
uvicorn
langchain
qdrant-client
google-generativeai
perplexityai
python-dotenv
markdown
